requirements.txt
click==8.1.3
Flask==1.0.2
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Jinja2==2.10
MarkupSafe==1.1.0
python-decouple==3.6
python-dotenv==0.20.0
python-engineio==3.13.2
python-socketio==4.6.0
six==1.11.0
SQLAlchemy==1.4.39
Werkzeug==0.14.1
WTForms==3.0.1
